# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Aviary API
  version: 0.1.0
  summary: Taxonomy + entity graph service with path-based data resolution.
  description: |
    Aviary is a schema-flexible entity store. Users declare **taxonomies**
    (schemas) at runtime, each with typed **fields** and **relationships** to
    other taxonomies. **Entities** are records whose `attributes` are a flat
    map of scalar values keyed by their taxonomy's field keys.

    Relationships are **computed, not stored**: `to_one` / `to_many` resolve
    at read time by matching `source_field` values on one side to
    `target_field` values on the other; `to_many_through` composes an
    ordered chain of other relationships. There are no foreign keys between
    entities in the database — every traversal re-resolves.

    ### Contract invariants
    - Non-2xx responses use the error envelope `{ error: { code, message, details? } }`.
    - Collections return ascending `id` order. Related entities in `to_many`
      traversal sort by target `id` ascending.
    - `POST /resolve` **always** returns 200. Per-path failures appear in
      `response.errors[path]`, never in an envelope.
    - Archived related entities are treated as missing during traversal.
      Archived **root** taxonomies and entities are still served on direct
      lookup — this is intentional.
    - Attributes are scalar only (string / integer / float / boolean /
      date / datetime / null). Objects and arrays are rejected.

...
  contact:
...
servers:
...
paths:
...
  /taxonomies:
    get:
      tags: [taxonomies]
      summary: List taxonomies.
      description: |
        Returns all taxonomies in ascending `id` order. Archived taxonomies
        are excluded unless `include_archived=true`.
      operationId: listTaxonomies
      parameters:
        - $ref: "#/components/parameters/IncludeArchived"
      responses:
        "200":
          description: Taxonomy collection.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaxonomyList" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      tags: [taxonomies]
      summary: Create a taxonomy.
      description: |
        Validates structure (unique keys, supported types, valid relationship
        refs including `to_many_through` cycle detection) and references
        before writing. `archived` defaults to `false`.

        Returns `409 conflict` when a taxonomy with the same `id` already
        exists.
      operationId: createTaxonomy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTaxonomy" }
            examples:
              patients: { $ref: "#/components/examples/PatientsTaxonomy" }
      responses:
        "201":
          description: Taxonomy created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Taxonomy" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /taxonomies/{taxonomy_id}:
    parameters:
      - $ref: "#/components/parameters/TaxonomyId"
    get:
      tags: [taxonomies]
      summary: Get a taxonomy by id.
      description: |
        Archived taxonomies are returned on direct lookup (archive status
        only affects list endpoints and traversal).
      operationId: getTaxonomy
      responses:
        "200":
          description: Taxonomy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Taxonomy" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    patch:
      tags: [taxonomies]
      summary: Update a taxonomy.
      description: |
        Allowed top-level keys: `name`, `archived`, `fields`, `relationships`.
        Any other key (including `id`) produces a `400 validation_error`.

        Provided keys **replace** the existing value (no merging of field or
        relationship arrays). When `fields` changes, the server runs an
        entity-compatibility guard: removing a field that entities use,
        changing a type that orphans values, or promoting an optional field
        to required without populating it returns `409 conflict` with a
        per-entity breakage map in `details.entities`.
      operationId: patchTaxonomy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchTaxonomy" }
            examples:
              rename:
                value:
                  name: Patients (v2)
              archive:
                value:
                  archived: true
      responses:
        "200":
          description: Taxonomy updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Taxonomy" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [taxonomies]
      summary: Delete a taxonomy.
      description: |
        Returns `409 conflict` if any entities still belong to the taxonomy
        or any other taxonomy still references it in a relationship.
      operationId: deleteTaxonomy
      responses:
        "204":
          description: Taxonomy deleted.
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /taxonomies/{taxonomy_id}/relationship-graph:
    parameters:
      - $ref: "#/components/parameters/TaxonomyId"
    get:
      tags: [taxonomies]
      summary: Introspect a taxonomy's relationship graph.
      description: |
        Returns a nested schema view rooted at the given taxonomy. Each node
        includes the root `taxonomy_id` key plus `fieldKey: fieldType` entries
        for every field, and nested nodes for each relationship.

        **Depth semantics** — `depth=1` means only the root's own fields;
        `depth=2` includes one relationship hop; `depth=3` includes two; and
        so on. Archived taxonomies are omitted from recursion. Cycle handling
        is automatic.

        Returns `413 response_too_large` when the produced graph exceeds the
        server's node budget.
      operationId: getRelationshipGraph
      parameters:
        - name: depth
          in: query
          description: Traversal depth (>= 1). Defaults to 2.
          required: false
          schema:
            type: integer
            minimum: 1
            default: 2
      responses:
        "200":
          description: Relationship graph.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RelationshipGraphResponse" }
              examples:
                patients:
                  value:
                    taxonomy_id: patients
                    depth: 2
                    graph:
                      taxonomy_id: patients
                      patient_number: string
                      first_name: string
                      last_name: string
                      care_team:
                        taxonomy_id: care_teams
                        team_code: string
                        assigned_nurse: string
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/ResponseTooLarge" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  parameters:
    TaxonomyId:
      name: taxonomy_id
      in: path
      required: true
      schema: { type: string, minLength: 1 }
      description: Taxonomy identifier.
...
    IncludeArchived:
      name: include_archived
      in: query
      required: false
      description: When `true`, include archived records. Defaults to `false`.
      schema:
        type: boolean
        default: false
  responses:
    ValidationError:
      description: Input failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            field:
              value:
                error:
                  code: validation_error
                  message: "attributes.date_of_birth: expected date (YYYY-MM-DD)"
                  details:
                    fields:
                      "attributes.date_of_birth": expected date
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            entity:
              value:
                error:
                  code: not_found
                  message: "Entity 'patient-9999' not found"
                  details:
                    resource: Entity
                    id: patient-9999
    Conflict:
      description: |
        Request is well-formed but the current state forbids it. Examples:
        duplicate id, taxonomy deletion blocked by entities or inbound refs,
        PATCH taxonomy incompatible with existing entities, or `ambiguous_to_one`
        during `/data` traversal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            duplicate:
              value:
                error:
                  code: conflict
                  message: "Taxonomy 'patients' already exists"
                  details:
                    id: patients
            incompatible:
              value:
                error:
                  code: conflict
                  message: "field change would orphan existing entity values"
                  details:
                    entities:
                      "patient-1001":
                        - removed field 'legacy_code' still in use
    ResponseTooLarge:
      description: Produced response would exceed server budgets.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            graph:
              value:
                error:
                  code: response_too_large
                  message: "relationship graph exceeded node budget"
    InternalError:
      description: Unhandled server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            internal:
              value:
                error:
                  code: internal_error
                  message: "internal server error"
  examples:
    PatientsTaxonomy:
      value:
        id: patients
        name: Patients
        archived: false
        fields:
          - { key: patient_number, type: string,   required: true,  is_key: true  }
          - { key: first_name,     type: string,   required: true,  is_key: false }
          - { key: last_name,      type: string,   required: true,  is_key: false }
          - { key: date_of_birth,  type: date,     required: true,  is_key: false }
          - { key: care_team_code, type: string,   required: false, is_key: false }
        relationships:
          - key: care_team
            target_taxonomy_id: care_teams
            cardinality: to_one
            match:
              - { source_field: care_team_code, target_field: team_code }
          - key: coaching_sessions
            target_taxonomy_id: coaching_sessions
            cardinality: to_many_through
            through: [current_enrollment, coaching_sessions]
...
  schemas:
...
    FieldType:
      type: string
      enum: [string, integer, boolean, float, date, datetime]
      description: Supported scalar field types.
    ErrorCode:
      type: string
      description: |
        Stable `code` values emitted on non-2xx envelopes. Clients may
        branch on these.
      enum:
        - validation_error
        - not_found
        - conflict
        - response_too_large
        - not_implemented
        - internal_error
        - unavailable
...
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              description: Optional structured detail about the failure.

    # ----- Taxonomy ----------------------------------------------------------
    TaxonomyField:
      type: object
      required: [key, type, required, is_key]
      additionalProperties: false
      properties:
        key:
          type: string
          minLength: 1
          description: Attribute name on entities.
        type: { $ref: "#/components/schemas/FieldType" }
        required:
          type: boolean
          description: When true, entities must provide this key on create.
        is_key:
          type: boolean
          description: |
            Marks the taxonomy's canonical identifier. Informational only —
            the primary uniqueness guarantee is the entity `id`.
    RelationshipMatch:
      type: object
      required: [source_field, target_field]
      additionalProperties: false
      properties:
        source_field: { type: string, minLength: 1 }
        target_field: { type: string, minLength: 1 }
    DirectRelationship:
      type: object
      required: [key, target_taxonomy_id, cardinality, match]
      additionalProperties: false
      properties:
        key: { type: string, minLength: 1 }
        target_taxonomy_id: { type: string, minLength: 1 }
        cardinality:
          type: string
          enum: [to_one, to_many]
        match:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/RelationshipMatch" }
    ThroughRelationship:
      type: object
      required: [key, target_taxonomy_id, cardinality, through]
      additionalProperties: false
      properties:
        key: { type: string, minLength: 1 }
        target_taxonomy_id: { type: string, minLength: 1 }
        cardinality:
          type: string
          enum: [to_many_through]
        through:
          type: array
          minItems: 1
          description: |
            Ordered list of relationship keys to walk from the source taxonomy
            to reach `target_taxonomy_id`. Composes existing relationships;
            does not declare its own `match`.
          items: { type: string, minLength: 1 }
    Relationship:
      description: |
        Discriminated on `cardinality`. `to_one` and `to_many` use `match`;
        `to_many_through` uses `through`.
      oneOf:
        - $ref: "#/components/schemas/DirectRelationship"
        - $ref: "#/components/schemas/ThroughRelationship"
      discriminator:
        propertyName: cardinality
        mapping:
          to_one: "#/components/schemas/DirectRelationship"
          to_many: "#/components/schemas/DirectRelationship"
          to_many_through: "#/components/schemas/ThroughRelationship"
    Taxonomy:
      type: object
      required: [id, name, archived, fields, relationships]
      additionalProperties: false
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        archived: { type: boolean }
        fields:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyField" }
        relationships:
          type: array
          items: { $ref: "#/components/schemas/Relationship" }
    CreateTaxonomy:
      type: object
      required: [id, name, fields, relationships]
      additionalProperties: false
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        archived:
          type: boolean
          default: false
        fields:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyField" }
        relationships:
          type: array
          items: { $ref: "#/components/schemas/Relationship" }
    PatchTaxonomy:
      type: object
      additionalProperties: false
      description: |
        Only `name`, `archived`, `fields`, `relationships` are allowed. Any
        other top-level key (including `id`) returns `400 validation_error`.
        `fields` and `relationships` REPLACE existing arrays; there is no
        merge.
      properties:
        name: { type: string, minLength: 1 }
        archived: { type: boolean }
        fields:
          type: array
          items: { $ref: "#/components/schemas/TaxonomyField" }
        relationships:
          type: array
          items: { $ref: "#/components/schemas/Relationship" }
    TaxonomyList:
      type: object
      required: [data]
      additionalProperties: false
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Taxonomy" }

    # ----- Entity ------------------------------------------------------------
...
    RelationshipGraphNode:
      type: object
      description: |
        A node in the taxonomy relationship graph. Every node carries a
        `taxonomy_id` key. Field entries are `<fieldKey>: <FieldType>` strings.
        Nested relationship entries are objects of the same shape.
      required: [taxonomy_id]
      properties:
        taxonomy_id: { type: string }
      additionalProperties:
        oneOf:
          - { $ref: "#/components/schemas/FieldType" }
          - { $ref: "#/components/schemas/RelationshipGraphNode" }
    RelationshipGraphResponse:
      type: object
      required: [taxonomy_id, depth, graph]
      additionalProperties: false
      properties:
        taxonomy_id: { type: string }
        depth:
          type: integer
          minimum: 1
        graph: { $ref: "#/components/schemas/RelationshipGraphNode" }

    # ----- Entity /data traversal -------------------------------------------
...